CORS 설정 변경, env파일 수정
