--- .env
@@ -1,1 +1,0 @@
-VUE_APP_API_URL=http://localhost:8087
