+++ .env.production
@@ -1,1 +1,1 @@
-VUE_APP_API_URL=http://183.99.3.15:8087
+VUE_APP_API_BASE_URL=http://183.99.3.15:8087
