+++ .gitignore
@@ -6,6 +6,8 @@ node_modules
 # local env files
 .env.local
 .env.*.local
+.env.development
+.env.production
 
 # Log files
 npm-debug.log*
