+++ src/api.js
@@ -1,0 +1,30 @@
+import axios from "axios";
+
+const api = axios.create({
+  baseURL: process.env.VUE_APP_API_BASE_URL,
+  headers: {
+    "Content-Type": "application/json",
+  },
+  withCredentials: true,
+});
+
+// 요청 인터셉터 추가
+axios.interceptors.request.use(
+  (config) => {
+    // localStorage에서 토큰 가져오기
+    const token = localStorage.getItem("token");
+    // 토큰이 존재하면 헤더에 추가
+
+    if (token) {
+      config.headers.Authorization = `Bearer ${token}`;
+    }
+    return config;
+  },
+  (error) => {
+    return Promise.reject(error);
+  }
+);
+
+console.log(process.env.VUE_APP_API_BASE_URL);
+
+export default api;
