+++ src/components/team/TeamRegister.vue
@@ -5,14 +5,14 @@
 
 <script setup lang="ts">
 import { defineEmits } from "vue";
-import axios from "axios";
+import api from "@/api";
 
 const emit = defineEmits(["team-registered"]);
 
 const registerTeam = async () => {
   try {
     // 팀 등록 API 호출
-    await axios.post("http://localhost:8087/teams", {
+    await api.post("/teams", {
       /* 팀 등록에 필요한 데이터 */
     });
     // 성공적으로 팀이 등록되면 부모 컴포넌트에 알림
