+++ src/views/AppLogin.vue
@@ -150,7 +150,7 @@
 </template>
 
 <script>
-import axios from "axios";
+import api from "@/api";
 import { ref } from "vue";
 import { useRouter } from "vue-router";
 
@@ -163,7 +163,7 @@ export default {
 
     async function login() {
       try {
-        const response = await axios.post("http://localhost:8087/auth/login", {
+        const response = await api.post("/auth/login", {
           username: userName.value,
           password: password.value,
         });
