+++ src/views/PostView.vue
@@ -142,7 +142,7 @@
 <script>
 import Breadcrumb from "../partials/AppBreadcrumb.vue";
 import AppTable from "../components/AppTable.vue";
-import axios from "axios";
+import api from "@/api";
 export default {
   name: "PostView",
   components: {
@@ -188,8 +188,8 @@ export default {
         params.keyword = this.searchQuery;
       }
 
-      axios
-        .get("http://localhost:8087/posts/search", { params })
+      api
+        .get("/posts/search", { params })
         .then((response) => {
           console.log("응답 데이터:", response.data); // 응답 데이터 로그 추가
           this.postData = response.data; // 응답 데이터를 postData에 저장
