+++ vue.config.js
@@ -1,3 +1,3 @@
 module.exports = {
-  transpileDependencies: []
-}
+  transpileDependencies: [],
+};
